Add document duplicate command
